requirements.txt
numpy==1.17.2
matplotlib==3.1.1
torchvision==0.4.2
torch==1.3.1
opencv-python==4.1.2.30
Keras==2.2.4
pandas==0.25.3
scipy==1.3.1
dlib==19.19.0
Pillow==7.0.0
scikit-learn==0.22.1
scikit-image==0.15.0
